+++ src/directives/theme-boxes.js
@@ -20,7 +20,7 @@
             template: '' +
                 '<div class="odswidget odswidget-theme-boxes">' +
                 '   <div ng-repeat="item in items" class="odswidget-theme-boxes__box" ods-facet-results="items" ods-facet-results-context="context" ods-facet-results-facet-name="theme">' +
-                '       <a ng-href="{{context.domainUrl}}/explore/?refine.theme={{encode(item.path)}}" target="_self" ods-tooltip="{{item.name}} ({{formatCount(item.count)}})" ods-tooltip-direction="bottom" style="display: block;">' +
+                '       <a ng-href="{{context.domainUrl}}/explore/?refine.theme={{encode(item.path)}}" target="_self" translate="ods-tooltip" translate-n="item.count" translate-plural="{{item.name}} ({{$count}} datasets)" ods-tooltip="{{item.name}} ({{$count}} dataset)" ods-tooltip-direction="bottom" style="display: block;">' +
                 '           <ods-theme-picto class="odswidget-theme-boxes__picto" theme="{{item.name}}"></ods-theme-picto>' +
                 '       </a>' +
                 '   </div>' +
@@ -28,15 +28,7 @@
             scope: {
                 context: '='
             },
-            controller: ['$scope', 'translate', function($scope, translate) {
-                $scope.formatCount = function(count) {
-                    // As it is very complicated to use ngPluralize with odsTooltip
-                    if (count > 1) {
-                        return count + ' ' + translate('datasets');
-                    } else {
-                        return count + ' ' + translate('dataset');
-                    }
-                };
+            controller: ['$scope', function($scope) {
                 $scope.encode = encodeURIComponent;
             }]
         };